requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
psycopg2-binary==2.9.9
pydantic>=2.0
pytest==8.0.0
psutil==5.9.6
kubernetes==29.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
numpy>=1.24.0
prophet>=1.1.0
ortools>=9.8
xgboost>=2.0.0
shap>=0.43.0
scikit-learn>=1.3.0
alembic>=1.13.0
SQLAlchemy>=2.0.0
slowapi>=0.1.9
python-barcode>=0.15.1
qrcode[pil]>=7.4
Pillow>=10.0.0
redis>=5.0.0
paho-mqtt>=2.0.0
python-multipart>=0.0.6
